css_string is now the same as a parserutils_dict_entry. This allows us to use dict entries directly as strings.
iChange the way in which selectors are represented. This significantly reduces memory requirements -- reducing the approximate usage count (excludes the string dictionary, which is about 360k) of allzengarden.css from 4,535,400 bytes to 2,414,312 bytes on a 64bit platform.

The string dictionary is now created and owned by the stylesheet object. The parser is just given access to this so that it can store strings in it.

svn path=/trunk/libcss/; revision=5809